refactor: move fs3i inpar to submodule
