+++ src/fs3i/4C_fs3i_input.cpp
@@ -5,7 +5,7 @@
 //
 // SPDX-License-Identifier: LGPL-3.0-or-later
 
-#include "4C_inpar_fs3i.hpp"
+#include "4C_fs3i_input.hpp"
 
 #include "4C_inpar_scatra.hpp"
 #include "4C_io_input_spec_builders.hpp"
@@ -13,7 +13,7 @@ FOUR_C_NAMESPACE_OPEN
 
 
 
-std::vector<Core::IO::InputSpec> Inpar::FS3I::valid_parameters()
+std::vector<Core::IO::InputSpec> FS3I::valid_parameters()
 {
   using namespace Core::IO::InputSpecBuilders;
