+++ src/global_legacy_module/CMakeLists.txt
@@ -21,6 +21,7 @@ set(_dependencies
     fluid
     fluid_ele
     fluid_xfluid
+    fs3i
     lubrication
     mat
     membrane
